requirements.txt
nltk==3.8.1
textblob
newspaper3k